refactor[#175] : 사용자 현재 위치를 찾는 커스텀 훅 생성
- 사용자의 현재 위치를 파악하면 MapStore에 저장
