+++ src/features/map/ui/MapControlButton.tsx
@@ -1,4 +1,3 @@
-import { useRef } from "react";
 import { useMap } from "@vis.gl/react-google-maps";
 import { FloatingButton } from "@/entities/map/ui";
 import { MapSnackbar } from "@/entities/map/ui/MapSnackbar";
@@ -10,6 +9,7 @@ import {
   MyLocationIcon,
   StarIcon,
 } from "@/shared/ui/icon";
+import { useCurrentLocation } from "../lib";
 import { useMapStore } from "../store";
 import { MarkingFormModal } from "./MarkingFormModal";
 
@@ -50,72 +50,18 @@ export const MyLocationButton = () => {
   const setIsCenterOnMyLocation = useMapStore(
     (state) => state.setIsCenterOnMyLocation,
   );
-  const setUserInfo = useMapStore((state) => state.setUserInfo);
 
-  const failureCount = useRef<number>(0);
-  const TIME_OUT = 1000;
+  const { setCurrentLocation } = useCurrentLocation(({ coords }) => {
+    const { latitude: lat, longitude: lng } = coords;
+
+    map.setCenter({ lat, lng });
+    setTimeout(() => {
+      setIsCenterOnMyLocation(true);
+    }, 0);
+  });
 
   const handleClick = () => {
-    const successCallback = ({ coords }: GeolocationPosition) => {
-      const { latitude: lat, longitude: lng } = coords;
-      map.setCenter({ lat, lng });
-      setUserInfo({
-        currentLocation: { lat, lng },
-        hasLocationPermission: true,
-      });
-      // GoogleMap의 CameraChanged 이벤트가 발생한 후에 시행 될 수 있도록
-      // setIsCenterOnMyLocation 을 이벤트 루프로 보내줍니다.
-      setTimeout(() => {
-        setIsCenterOnMyLocation(true);
-      }, 0);
-    };
-
-    const errorCallback = (error: GeolocationPositionError) => {
-      switch (error.code) {
-        case 1 /* PERMISSION_DENIED */:
-          setUserInfo({
-            currentLocation: { lat: 0, lng: 0 },
-            hasLocationPermission: false,
-          });
-          throw new Error(
-            "위치 제공을 허용 한 후 사용 가능한 기능입니다\n내 위치 제공을 허용해주세요",
-          );
-
-        case 2 /* POSITION_UNAVAILABLE */:
-          if (failureCount.current >= 3) {
-            failureCount.current = 0;
-            throw new Error(
-              "위치 정보를 가져오는데 문제가 발생했습니다\n잠시 후 다시 시도해주세요.",
-            );
-          }
-          window.navigator.geolocation.getCurrentPosition(
-            successCallback,
-            errorCallback,
-            {
-              ...options,
-              timeout: TIME_OUT + 2 ** failureCount.current * 100,
-              enableHighAccuracy: false,
-            },
-          );
-          break;
-        case 3 /* TIMEOUT */:
-          throw new Error(
-            "위치 정보를 가져오는데 시간이 너무 오래 걸립니다\n잠시 후 다시 시도해주세요.",
-          );
-      }
-    };
-
-    const options: PositionOptions = {
-      enableHighAccuracy: true,
-      timeout: TIME_OUT,
-      maximumAge: 0,
-    };
-
-    window.navigator.geolocation.getCurrentPosition(
-      successCallback,
-      errorCallback,
-      options,
-    );
+    setCurrentLocation();
   };
 
   return (
